# swagger/openapi.yaml
openapi: 3.0.3
info:
  title: ASAAS Integration Test API
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /customers:
    post:
      summary: Cria um cliente no Asaas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
      responses:
        '201':
          description: Cliente criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
    get:
      summary: Busca um cliente direto no Asaas
      parameters:
        - in: query
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Cliente encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
  /payments:
    post:
      summary: Cria um pagamento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
      responses:
        '201':
          description: Pagamento criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
    get:
      summary: Consulta pagamento direto no Asaas
      parameters:
        - in: query
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Pagamento encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
  /subscriptions:
    post:
      summary: Cria uma assinatura
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionRequest'
      responses:
        '201':
          description: Assinatura criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
  /subscriptions/cancel:
    post:
      summary: Cancela uma assinatura
      parameters:
        - in: query
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Cancelada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
  /invoices:
    post:
      summary: Cria uma nota fiscal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceRequest'
      responses:
        '201':
          description: Nota criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponse'
    get:
      summary: Consulta nota fiscal direto no Asaas
      parameters:
        - in: query
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Nota fiscal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponse'
  /webhooks/asaas:
    post:
      summary: Recebe webhooks do Asaas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookEvent'
      responses:
        '204':
          description: Evento aceito
components:
  schemas:
    CustomerRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        email:
          type: string
        cpfCnpj:
          type: string
        phone:
          type: string
        mobilePhone:
          type: string
        address:
          type: string
        addressNumber:
          type: string
        complement:
          type: string
        province:
          type: string
        postalCode:
          type: string
        notificationDisabled:
          type: boolean
        additionalEmails:
          type: string
    CustomerResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        externalReference:
          type: string
    PaymentRequest:
      type: object
      required: [customer, billingType, value, dueDate, callback]
      properties:
        customer:
          type: string
          description: ID do cliente no seu sistema.
        billingType:
          type: string
        value:
          type: number
        dueDate:
          type: string
          format: date
        description:
          type: string
        installmentCount:
          type: integer
        callback:
          $ref: '#/components/schemas/PaymentCallback'
    PaymentResponse:
      type: object
      properties:
        id:
          type: string
        customer:
          type: string
        billingType:
          type: string
        value:
          type: number
        status:
          type: string
        externalReference:
          type: string
        invoiceUrl:
          type: string
        transactionReceiptUrl:
          type: string
    PaymentCallback:
      type: object
      required: [successUrl, autoRedirect]
      properties:
        successUrl:
          type: string
        autoRedirect:
          type: boolean
    SubscriptionRequest:
      type: object
      required: [customer, billingType, value, nextDueDate, cycle]
      properties:
        customer:
          type: string
          description: ID do cliente no seu sistema.
        billingType:
          type: string
        value:
          type: number
        nextDueDate:
          type: string
          format: date
        cycle:
          type: string
        description:
          type: string
        endDate:
          type: string
          format: date
        maxPayments:
          type: integer
    SubscriptionResponse:
      type: object
      properties:
        id:
          type: string
        customer:
          type: string
        status:
          type: string
        value:
          type: number
    InvoiceRequest:
      type: object
      required: [payment, serviceDescription, observations, value, deductions, effectiveDate, municipalServiceName, taxes]
      properties:
        payment:
          type: string
          description: ExternalReference da cobrança no seu sistema.
        serviceDescription:
          type: string
        observations:
          type: string
        externalReference:
          type: string
        value:
          type: number
        deductions:
          type: number
        effectiveDate:
          type: string
          format: date
        municipalServiceId:
          type: string
        municipalServiceCode:
          type: string
        municipalServiceName:
          type: string
        updatePayment:
          type: boolean
        taxes:
          $ref: '#/components/schemas/InvoiceTaxes'
    InvoiceResponse:
      type: object
      properties:
        id:
          type: string
        customer:
          type: string
        status:
          type: string
        value:
          type: number
        externalReference:
          type: string
        paymentLink:
          type: string
    InvoiceTaxes:
      type: object
      required: [retainIss, cofins, csll, inss, ir, pis, iss]
      properties:
        retainIss:
          type: boolean
        cofins:
          type: number
        csll:
          type: number
        inss:
          type: number
        ir:
          type: number
        pis:
          type: number
        iss:
          type: number
    WebhookEvent:
      type: object
      properties:
        event:
          type: string
        payment:
          $ref: '#/components/schemas/PaymentResponse'
        invoice:
          $ref: '#/components/schemas/InvoiceResponse'
        subscription:
          $ref: '#/components/schemas/SubscriptionResponse'